Add `GameParams` and `GameMetadata` classes
`Game` class is overly complex and contains attributes with tree
different responsibilities:

- store user-defined parameters (dataset name, difficulty level, precision mode and so on)
- store internal metadata (game name, players, creation date)
- store internal game data (places to guess, current run, previous results from players, and so on)

Parameters and metadata must be copied when relaunching a game, whereas game data is game-specific and must be cleared when relaunching.
