+++ amstramdam/events/user.py
@@ -64,14 +64,11 @@ def change_game(data: PartialGameParams) -> None:
     params = merge_params(data)
     print(*[f"{k}={v}" for k, v in params.items()], sep=", ")
     new_game_name, game = manager.create_game(
-        n_run=params["runs"],
-        duration=params["duration"],
-        difficulty=params["difficulty"],
+        dataset_name=params["map"],
+        level=params["difficulty"],
         is_public=params["public"],
         precision_mode=params["precision_mode"],
-        allow_zoom=params["zoom"],
-        map=params["map"],
-        wait_time=params["wait_time"],
+        wait_duration=params["wait_time"],
     )
     url = url_for("serve_game", name=new_game_name)
     print(url)
