+++ requirements-full.txt
@@ -1,40 +1,136 @@
-bidict==0.22.1
-black==23.3.0
-blinker==1.6.2
-click==8.1.3
-dnspython==2.3.0
-eventlet==0.33.3
-flake8==6.0.0
-Flask==2.3.2
-Flask-SocketIO==5.3.4
-flask-talisman==1.0.0
-geographiclib==2.0
-geopy==2.3.0
-greenlet==2.0.2
-iniconfig==2.0.0
-itsdangerous==2.1.2
-Jinja2==3.1.2
-lxml==4.9.2
-MarkupSafe==2.1.2
-mccabe==0.7.0
-mypy==1.3.0
-mypy-extensions==1.0.0
-numpy==1.24.3
-packaging==23.1
-pandas==2.0.1
-pathspec==0.11.1
-platformdirs==3.5.1
+annotated-types==0.6.0
+anyio==4.3.0
+argon2-cffi==23.1.0
+argon2-cffi-bindings==21.2.0
+arrow==1.3.0
+asttokens==2.4.1
+async-lru==2.0.4
+atomicwrites==1.4.0
+attrs==23.2.0
+Babel==2.15.0
+beautifulsoup4==4.12.3
+bidict==0.21.4
+black==22.1.0
+bleach==6.1.0
+certifi==2021.10.8
+cffi==1.16.0
+charset-normalizer==3.3.2
+click==8.0.4
+click-plugins==1.1.1
+cligj==0.7.2
+colorama==0.4.4
+comm==0.2.2
+debugpy==1.8.1
+decorator==5.1.1
+defusedxml==0.7.1
+dnspython==2.2.0
+eventlet==0.33.0
+exceptiongroup==1.2.1
+executing==2.0.1
+fastjsonschema==2.19.1
+flake8==4.0.1
+Flask==2.0.3
+Flask-SocketIO==5.1.1
+flask-talisman==0.8.1
+fqdn==1.5.1
+geographiclib==1.52
+geojson==2.5.0
+geopy==2.2.0
+greenlet==1.1.2
+h11==0.14.0
+httpcore==1.0.5
+httpx==0.27.0
+idna==3.7
+iniconfig==1.1.1
+ipykernel==6.29.4
+ipython==8.24.0
+isoduration==20.11.0
+itsdangerous==2.1.0
+jedi==0.19.1
+Jinja2==3.0.3
+json5==0.9.25
+jsonpointer==2.4
+jsonschema==4.22.0
+jsonschema-specifications==2023.12.1
+jupyter-client==8.6.1
+jupyter-core==5.7.2
+jupyter-events==0.10.0
+jupyter-lsp==2.2.5
+jupyter-server==2.14.0
+jupyter-server-terminals==0.5.3
+jupyterlab-pygments==0.3.0
+jupyterlab-server==2.27.1
+lxml==4.8.0
+MarkupSafe==2.1.0
+matplotlib-inline==0.1.7
+mccabe==0.6.1
+mistune==3.0.2
+munch==2.5.0
+mypy==0.931
+mypy-extensions==0.4.3
+nbclient==0.10.0
+nbconvert==7.16.4
+nbformat==5.10.4
+nest-asyncio==1.6.0
+notebook-shim==0.2.4
+numpy==1.22.2
+overrides==7.7.0
+packaging==24.0
+pandas==1.4.1
+pandocfilters==1.5.1
+parso==0.8.4
+pathspec==0.9.0
+platformdirs==2.5.0
 pluggy==1.0.0
-pycodestyle==2.10.0
-pyflakes==3.0.1
-pytest==7.3.1
+prometheus-client==0.20.0
+prompt-toolkit==3.0.43
+psutil==5.9.8
+pure-eval==0.2.2
+py==1.11.0
+pycodestyle==2.8.0
+pycparser==2.22
+pydantic==2.7.1
+pydantic-core==2.18.2
+pydantic-settings==2.2.1
+pyflakes==2.4.0
+pygments==2.18.0
+pyparsing==3.0.7
+pyproj==3.3.0
+pytest==7.0.1
 python-dateutil==2.8.2
-python-engineio==4.4.1
-python-socketio==5.8.0
-pytz==2023.3
+python-dotenv==1.0.1
+python-engineio==4.3.1
+python-json-logger==2.0.7
+python-socketio==5.5.2
+pytz==2021.3
+pywin32==306
+pywinpty==2.0.13
+PyYAML==6.0.1
+pyzmq==26.0.3
+referencing==0.35.1
+requests==2.31.0
+rfc3339-validator==0.1.4
+rfc3986-validator==0.1.1
+rpds-py==0.18.1
+Send2Trash==1.8.3
+Shapely==1.8.1.post1
 six==1.16.0
-tqdm==4.65.0
-typing_extensions==4.6.2
-tzdata==2023.3
-Unidecode==1.3.6
-Werkzeug==2.3.4
+sniffio==1.3.1
+soupsieve==2.5
+stack-data==0.6.3
+terminado==0.18.1
+tinycss2==1.3.0
+tomli==2.0.1
+tornado==6.4
+tqdm==4.62.3
+traitlets==5.14.3
+types-python-dateutil==2.9.0.20240316
+typing-extensions==4.11.0
+Unidecode==1.3.3
+uri-template==1.3.0
+urllib3==2.2.1
+wcwidth==0.2.13
+webcolors==1.13
+webencodings==0.5.1
+websocket-client==1.8.0
+Werkzeug==2.0.3
