+++ server.py
@@ -29,13 +29,12 @@ if __name__ == "__main__":
         message = "Live-reloading enabled"
         if debug_game:
             name, game = manager.create_game(
-                n_run=3,
+                dataset_name="paris_subway",
+                n_runs=3,
                 is_permanent=True,
                 duration=5,
                 map="paris_subway",
-                wait_time=4,
-                is_public=True,
-                allow_zoom=True,
+                wait_duration=4,
                 force_name="__debug__",
             )
             for _ in range(3):
